deprecations and moving files
