+++ pymc_marketing/mmm/__init__.py
@@ -13,7 +13,7 @@
 #   limitations under the License.
 """Marketing Mix Models (MMM)."""
 
-from pymc_marketing.mmm import base, delayed_saturated_mmm, preprocessing, validating
+from pymc_marketing.mmm import base, mmm, preprocessing, validating
 from pymc_marketing.mmm.base import BaseValidateMMM, MMMModelBuilder
 from pymc_marketing.mmm.components.adstock import (
     AdstockTransformation,
@@ -37,8 +37,8 @@ from pymc_marketing.mmm.components.saturation import (
     register_saturation_transformation,
     saturation_from_dict,
 )
-from pymc_marketing.mmm.delayed_saturated_mmm import MMM, DelayedSaturatedMMM
 from pymc_marketing.mmm.fourier import MonthlyFourier, YearlyFourier
+from pymc_marketing.mmm.mmm import MMM
 from pymc_marketing.mmm.preprocessing import (
     preprocessing_method_X,
     preprocessing_method_y,
@@ -49,7 +49,6 @@ __all__ = [
     "AdstockTransformation",
     "BaseValidateMMM",
     "DelayedAdstock",
-    "DelayedSaturatedMMM",
     "GeometricAdstock",
     "HillSaturation",
     "HillSaturationSigmoid",
@@ -71,7 +70,7 @@ __all__ = [
     "register_adstock_transformation",
     "YearlyFourier",
     "base",
-    "delayed_saturated_mmm",
+    "mmm",
     "preprocessing",
     "preprocessing_method_X",
     "preprocessing_method_y",
